// src/app/shared/constants/enums/core/default.enum.ts
/**
 * Default Constant enum for default features
 */
export enum DefaultConstant {
  TIMEZONE = 'Asia/Kolkata',
  SYSTEM_DATETIME_FORMAT = 'YYYY-MM-DDTHH:mm:ss',
  LOCAL_DATETIME_FORMAT = 'YYYY-MM-DD HH:mm:ss',
  LOCAL_DATE_FORMAT = 'D-M-YYYY',
  TIME_12_FORMAT = 'hh:mm:ss a',
  TIME_24_FORMAT = 'HH:mm:ss',
  USER_IMAGE = 'default.jpg',
  DEFAULT_FILE_RESPONSE = 'data.fullPath',
}